python/get_google_cloud_services/requirements.txt
google-cloud-resource-manager
google-cloud-service-usage